backend/pyproject.toml
[project]
name = "mentor-backend"
version = "0.2.0"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.21",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.6",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "unstructured[pdf,docx,md]>=0.15.0",
    "opencv-python-headless>=4.8.0",
    "lingua-language-detector>=2.0.0",
    "tree-sitter==0.20.4",
    "tree-sitter-languages==1.10.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "fpdf2>=2.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app", "tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
